Fix my divide by zero bug that has been infrequently crashing my agents.
I think this only occurred if the number of local steps taken was an
exact multiple of the logging interval (503) AND we happen to be exactly
at the end of an episode. Each of these colliding occurred with very low
probability, but when running for millions of steps, it did occur, and
crash the agent.